JetBrains: Cody: Refactor: Cleanup language util functions
