+++ client/jetbrains/src/main/kotlin/com/sourcegraph/utils/CodyLanguageUtil.kt
@@ -8,22 +8,11 @@ import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.project.Project
 import com.sourcegraph.config.ConfigUtil
 
-class CodyLanguageUtil {
-    companion object {
-        fun getLanguage(project: Project, document: Document): Language? {
-            return LanguageUtil.getLanguageForPsi(
-                project,
-                FileDocumentManager.getInstance().getFile(document))
-        }
-
-        fun getLanguage(editor: Editor): Language? {
-            val project = editor.project ?: return null
-            return getLanguage(project, editor.document)
-        }
-
-        fun isLanguageBlacklisted(editor: Editor): Boolean {
-            val language = getLanguage(editor) ?: return false
-            return ConfigUtil.getBlacklistedAutocompleteLanguageIds().contains(language.id)
-        }
+object CodyLanguageUtil {
+    @JvmStatic
+    fun getLanguage(project: Project, document: Document): Language? {
+        return LanguageUtil.getLanguageForPsi(
+            project,
+            FileDocumentManager.getInstance().getFile(document))
     }
 }
